refactor(codex): separate hook identity model
